Refactored pool, removed queryString, searchRadius from schema
It would make more sense to not store the queryString and searchRadius
in db, since it heavily depends on the output of the google nearby
search api and the user satisfaction. User would certainly like to try
out different queryStrings, and searchRadius to ensure he/she gets the
best result for that particular place.

Further Changes:

- Added doc strings to pool service
- Directly destructured the params in the function signature
- added internal service _updateCentroidInPlace which takes in the pool
object and updates the centroid inplace